mlt-templates/tf-dist-mnist-gpu/requirements.txt
tensorflow-gpu
pycodestyle==2.3.1
scipy==1.0.1
